Enable password set in case of rescued instances
In the case of rescued instances, the user password has already
been set for an instance, and it cannot be posted anymore to the
http metadata service. In this case we still have to set the new
password and warn the user that the password has already been
posted to the metadata service and it cannot be changed anymore.

Change-Id: Iaa8f34d52e2b00ac2dea88a401c83e45899b4f54
Closes-Bug: #1391818