Adds support for MIME base64 encoded parts in the userdata
According to [1], Content-Transfer-Encoding can have one of the
following values (case insensitive): "BASE64" / "QUOTED-PRINTABLE" /
"8BIT"   / "7BIT" / "BINARY" / x-token.
Values "8bit", "7bit", and "binary" all imply that NO encoding has been
performed.
"x-token" refers to a custom implementation that can be done, where the
encoding agent and the decoding agent must agree upon. This case is not
addressed by the patch.
"QUOTED-PRINTABLE" is an encoding not addressed by this patch, as there
is no cloud provider that encodes in this format.
"base64" encoding is addressed by this patch, and the decoding is
performed for all the multipart content types currently supported.

[1]https://www.w3.org/Protocols/rfc1341/5_Content-Transfer-Encoding.html

Change-Id: Iac29f3a287bd478f7f0b55d9911ae47f3a5890fb
Closes-Bug: #1696420